requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Pydantic
pydantic==2.5.0

# Google AI (Gemini)
google-generativeai==0.3.1

# Google APIs (Calendar, Gmail)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Selenium (Optional - for advanced browser features)
selenium==4.16.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0

# CORS
fastapi-cors==0.0.6